Save and show recently used characters
This saves the chosen character sequence in a file (roll-appending the
older ones) and offers them first.
It is quite stupid, not showing their description and only using the
last recently used ones instead of a frecency measure.
This makes it quite easy to implement and circumvents other problems:
What is the correct description for several characters? (We don't show a
description at all.) Should it remove the duplicate from the data files?
(Since we need one entry with the description, no.) How should we encode
the frecency? (Not at all, LRU is enough.)

Issue: #39